+++ extension/src/XliffEditor/HTML.ts
@@ -1,9 +1,12 @@
+import { isNullOrUndefined } from "util";
+
 export function checkbox(a: HTMLAttributes): string {
-    return `<input type="checkbox" ${a.id ? 'id="' + a.id + '"' : ''} ${a.name ? 'name="' + a.name + '"' : ''} ${a.checked ? " checked " : ""} ${a.disabled ? " disabled " : ""}>`;
+    a.type = "checkbox";
+    return `<input ${attributeString(a)}>`;
 }
 
 export function table(a: HTMLAttributes, columns: HTMLTag[]): string {
-    return `<table ${a.id ? 'id="' + a.id + '"' : ''}>${tr({}, columns)}</table>`;
+    return `<table ${attributeString(a)}>${tr({}, columns)}</table>`;
 }
 
 export function tableHeader(headers: string[]): string {
@@ -16,11 +19,11 @@ export function tableHeader(headers: string[]): string {
 }
 
 function th(a: HTMLAttributes, content: string): string {
-    return `<th class="${a.class}">${content}</th>`;
+    return `<th ${attributeString(a)}>${content}</th>`;
 }
 
 export function tr(a: HTMLAttributes, columns: HTMLTag[]): string {
-    let row: string = `<tr ${a.id ? 'id="' + a.id + '"' : ''}>`;
+    let row: string = `<tr ${attributeString(a)}>`;
     columns.forEach(c => {
         row += td(c);
     });
@@ -29,25 +32,26 @@ export function tr(a: HTMLAttributes, columns: HTMLTag[]): string {
 }
 
 function td(param: HTMLTag): string {
-    return `<td ${param.a?.align ? 'align="' + param.a.align + '"' : ''}>${param.content}</td>`;
+    return `<td ${attributeString(param.a)}>${param.content}</td>`;
 }
 
 export function div(a: HTMLAttributes, content: string): string {
-    let _div: string = `<div ${a.id ? 'id="' + a.id + '"' : ''} ${a.class ? 'class="' + a.class + '"' : ''} ${a.name ? 'name="' + a.name + '"' : ''}>`;
+    let _div: string = `<div ${attributeString(a)}>`;
     _div += content;
     _div += "</div>";
     return _div;
 }
 
 export function textArea(a: HTMLAttributes, content: string): string {
-    let tarea: string = `<textarea ${a.id ? 'id="' + a.id + '"' : ''} ${a.class ? 'class="' + a.class + '"' : ''} ${a.name ? 'name="' + a.name + '"' : ''} ${a.type ? 'type="' + a.type + '"' : ''}>`;
+    let tarea: string = `<textarea ${attributeString(a)}>`;
     tarea += content;
     tarea += "</textarea>";
     return tarea;
 }
 
 export function button(a: HTMLAttributes, content: string): string {
-    let btn: string = `<button ${a.id ? 'id="' + a.id + '"' : ''} ${a.class ? 'class="' + a.class + '"' : ''} ${a.onClick ? 'onClick="' + a.onClick + '"' : ''}>`;
+    // let btn: string = `<button ${a.id ? 'id="' + a.id + '"' : ''} ${a.class ? 'class="' + a.class + '"' : ''} ${a.onClick ? 'onClick="' + a.onClick + '"' : ''}>`;
+    let btn: string = `<button ${attributeString(a)}>`;
     btn += content;
     btn += "</button>";
     return btn;
@@ -57,7 +61,27 @@ export function br(noOfLinebreaks: number = 1): string {
     return (new Array<string>(noOfLinebreaks)).fill("<br/>").join("");
 }
 
-interface HTMLAttributes {
+export function attributeString(attributes?: HTMLAttributes): string {
+    let a = "";
+    if (!isNullOrUndefined(attributes)) {
+        Object.entries(attributes).forEach(attrib => {
+            switch (attrib[0]) {
+                case "checked":
+                    a += attrib[1] ? " checked" : "";
+                    break;
+                case "disabled":
+                    a += attrib[1] ? " disabled" : "";
+                    break;
+                default:
+                    a += ` ${attrib[0]}="${attrib[1]}"`;
+                    break;
+            }
+            a.trim();
+        });
+    }
+    return a.trim();
+}
+export interface HTMLAttributes {
     id?: string;
     class?: string;
     name?: string;