+++ extension/src/test/ExternalResources.test.ts
@@ -13,13 +13,14 @@ suite("External Resources Tests", function () {
     const fullUrl = 'https://nabaltools.file.core.windows.net/shared/base_app_lang_files/sv-se.json?sv=2019-12-12&ss=f&srt=o&sp=r&se=2021-11-25T05:28:10Z&st=2020-11-24T21:28:10Z&spr=https&sig=JP3RwQVCZBo16vJCznojVIMvPOHgnDuH937ppzPmEqQ%3D';
     const sasToken = 'sv=2019-12-12&ss=f&srt=o&sp=r&se=2021-11-25T05:28:10Z&st=2020-11-24T21:28:10Z&spr=https&sig=JP3RwQVCZBo16vJCznojVIMvPOHgnDuH937ppzPmEqQ%3D';
     const baseUrl = 'https://nabaltools.file.core.windows.net/shared/base_app_lang_files/';
+    const TIMEOUT = 30000;
 
     test("ExternalResource.get()", async function () {
         // Only run in GitHub Workflow
         if (!process.env.GITHUB_ACTION) {
             this.skip();
         }
-        this.timeout(10000);
+        this.timeout(TIMEOUT);
         const extResource = new ExternalResource('sv-se.json', href);
         const writeStream = createWriteStream(path.resolve(__dirname, "test.json"), "utf8");
         await extResource.get(writeStream);
@@ -41,7 +42,7 @@ suite("External Resources Tests", function () {
         if (!process.env.GITHUB_ACTION) {
             this.skip();
         }
-        this.timeout(5000);
+        this.timeout(TIMEOUT);
         const exportPath = path.resolve(__dirname);
         let blobContainer = new BlobContainer(exportPath, baseUrl, sasToken);
         blobContainer.addBlob('sv-se.json');